Restructured singleton-ness of sync_pim.py

+++ KitchenSyncD_mk2_desktop/sync_pim.py
@@ -6,8 +6,7 @@ import tempfile
 import shutil
 import sqlite3
 import hashlib
-from tendo import singleton
-me = singleton.SingleInstance()
+from pid import PidFile
 
 # functions
 def hashfile(afile, hasher, blocksize=65536):
@@ -42,37 +41,6 @@ def copyback():
     sshcom("am broadcast -a com.moosd.kitchensyncd.UPDATE_CALENDAR -n com.moosd.kitchensyncd/.CalendarChangeReceiver")
 #    unsetsyncprogress()
 
-# copy our files to a temp place
-dirpath = tempfile.mkdtemp()
-curdir = os.getcwd()
-shutil.copy("/srv/http/baikal/Specific/db/db.sqlite", dirpath)
-shutil.copy("/srv/http/baikal/Specific/db/db.sqlite.history", dirpath)
-
-# copy their files to a temp place
-os.chdir(dirpath)
-shutil.copy("db.sqlite", "db.sqlite.import")
-shutil.copy("db.sqlite.history", "db.sqlite.import.history")
-rsyncfroms("/sdcard/www/baikal/Specific/db/db.sqlite", "db.sqlite.import")
-rsyncfroms("/sdcard/www/baikal/Specific/db/db.sqlite.history", "db.sqlite.import.history")
-
-# if hashes are equal, copy back and exit...
-filea, fileb = (hashfile(open('db.sqlite', 'rb'), hashlib.sha256()), hashfile(open('db.sqlite.import', 'rb'), hashlib.sha256()))
-if filea == fileb:
-    copyback()
-    shutil.rmtree(dirpath)
-    sys.exit(0)
-
-# do sync...
-tables = ("calendarobjects", "cards")
-columns = { "calendarobjects": ["uri", "lastmodified", "calendardata", "calendarid", "etag", "size", "componenttype", "firstoccurence", "lastoccurence"], "cards": ["uri", "lastmodified", "addressbookid", "carddata"] }
-
-db = sqlite3.connect('db.sqlite')
-db_history = sqlite3.connect('db.sqlite.history')
-db_import = sqlite3.connect('db.sqlite.import')
-db_import_history = sqlite3.connect('db.sqlite.import.history')
-
-
-# first, update histories for both
 def updatehistory(file, file_hist):
     # go through file, if not present in history, mark as new
     file_hist.execute("create table if not exists entries(tab text not null, filename TEXT NOT NULL, timestamp integer, deleted integer not null);")
@@ -90,10 +58,6 @@ def updatehistory(file, file_hist):
             file_hist.execute("update entries set deleted=1 where tab=? and filename=?", (row[0], row[1]))
     pass
 
-updatehistory(db, db_history)
-updatehistory(db_import, db_import_history)
-
-# then add new objects to mine, then theirs
 def addnewobjects(frm, to):
     for t in tables:
         qstr = ",".join(columns[t])
@@ -110,10 +74,6 @@ def addnewobjects(frm, to):
                 else:
                     frm.execute("update %s set %s where uri='%s'" % (t, qqqstr, row[0]), ts)
 
-addnewobjects(db_import, db)
-addnewobjects(db, db_import)
-
-# then sync deletions
 def syncdeletionshistories(frm, to):
     # go through history file. If present in other history file, update self with older timestamp
     cursor = frm.execute("select filename, timestamp, deleted from entries")
@@ -127,19 +87,11 @@ def syncdeletionshistories(frm, to):
                 if otherrow[2] == 1:
                     frm.execute("update entries set deleted=?, timestamp=? where filename=?", (1, otherrow[1], row[0]))
 
-syncdeletionshistories(db_import_history, db_history)
-syncdeletionshistories(db_history, db_import_history)
-
 def syncdeletionsproper(file, file_hist):
     cursor = file_hist.execute("select tab, filename from entries where deleted=1")
     for row in cursor:
         file.execute("delete from %s where uri=?" % (row[0]), (row[1],))
 
-syncdeletionsproper(db, db_history)
-syncdeletionsproper(db_import, db_import_history)
-
-# update ctags if not equal
-ctagtables = ["calendars", "addressbooks"]
 def ctagupdate(file, file2):
     for t in ctagtables:
         cursor = file.execute("select id, ctag from %s" % t)
@@ -152,25 +104,76 @@ def ctagupdate(file, file2):
                     file.execute("update %s set ctag=? where id=?" % t, (ctag, row[0]))
                     file2.execute("update %s set ctag=? where id=?" % t, (ctag, row[0]))
 
-ctagupdate(db, db_import)
-#ctagupdate(db_import, db)
 
-db.commit()
-db_import.commit()
-db_history.commit()
-db_import_history.commit()
+ctagtables = ["calendars", "addressbooks"]
+tables = ("calendarobjects", "cards")
+columns = { "calendarobjects": ["uri", "lastmodified", "calendardata", "calendarid", "etag", "size", "componenttype", "firstoccurence", "lastoccurence"], "cards": ["uri", "lastmodified", "addressbookid", "carddata"] }
 
-db.close()
-db_history.close()
-db_import.close()
-db_import_history.close()
+# copy our files to a temp place
+with PidFile(piddir='/tmp'):
+    dirpath = tempfile.mkdtemp()
+    curdir = os.getcwd()
+    shutil.copy("/srv/http/baikal/Specific/db/db.sqlite", dirpath)
+    shutil.copy("/srv/http/baikal/Specific/db/db.sqlite.history", dirpath)
 
-# copy back
-copyback()
-shutil.copy("db.sqlite", "/srv/http/baikal/Specific/db/db.sqlite.backup")
-shutil.copy("db.sqlite", "/srv/http/baikal/Specific/db/db.sqlite")
-shutil.copy("db.sqlite.history", "/srv/http/baikal/Specific/db/db.sqlite.history")
+    # copy their files to a temp place
+    os.chdir(dirpath)
+    shutil.copy("db.sqlite", "db.sqlite.import")
+    shutil.copy("db.sqlite.history", "db.sqlite.import.history")
+    rsyncfroms("/sdcard/www/baikal/Specific/db/db.sqlite", "db.sqlite.import")
+    rsyncfroms("/sdcard/www/baikal/Specific/db/db.sqlite.history", "db.sqlite.import.history")
 
-# clean up
-#print(dirpath)
-shutil.rmtree(dirpath)
+    # if hashes are equal, copy back and exit...
+    filea, fileb = (hashfile(open('db.sqlite', 'rb'), hashlib.sha256()), hashfile(open('db.sqlite.import', 'rb'), hashlib.sha256()))
+    if filea == fileb:
+        copyback()
+        shutil.rmtree(dirpath)
+        sys.exit(0)
+
+    # do sync...
+    db = sqlite3.connect('db.sqlite')
+    db_history = sqlite3.connect('db.sqlite.history')
+    db_import = sqlite3.connect('db.sqlite.import')
+    db_import_history = sqlite3.connect('db.sqlite.import.history')
+
+    # first, update histories for both
+    updatehistory(db, db_history)
+    updatehistory(db_import, db_import_history)
+
+    # then add new objects to mine, then theirs
+    addnewobjects(db_import, db)
+    addnewobjects(db, db_import)
+
+    # then sync deletions
+    syncdeletionshistories(db_import_history, db_history)
+    syncdeletionshistories(db_history, db_import_history)
+
+    syncdeletionsproper(db, db_history)
+    syncdeletionsproper(db_import, db_import_history)
+
+    # update ctags if not equal
+    ctagupdate(db, db_import)
+    #ctagupdate(db_import, db)
+
+    db.commit()
+    db_import.commit()
+    db_history.commit()
+    db_import_history.commit()
+
+    db.close()
+    db_history.close()
+    db_import.close()
+    db_import_history.close()
+
+    # copy back
+    copyback()
+    shutil.copy("db.sqlite", "/srv/http/baikal/Specific/db/db.sqlite.backup")
+    shutil.copy("db.sqlite", "/srv/http/baikal/Specific/db/db.sqlite")
+    shutil.copy("db.sqlite.history", "/srv/http/baikal/Specific/db/db.sqlite.history")
+
+    # notify thunderbird
+
+
+    # clean up
+    #print(dirpath)
+    shutil.rmtree(dirpath)
